auth_service/requirements.txt
Flask
SQLAlchemy
PyMySQL
PyJWT